pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kaggle_template"
version = "0.1.0"
description = "Lightweight Kaggle competition starter template"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.25",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "joblib>=1.3",
]

[project.optional-dependencies]
viz = ["matplotlib>=3.7", "seaborn>=0.12"]
notebook = ["jupyter>=1.0.0"]
kaggle = ["kaggle>=1.6.0"]

[tool.kaggle-template]
default-config = "configs/base.yaml"
